Prevent data too long error on MySQL 5.7
Older mysql version would simply truncate. This change bumps the max
length to 2048 and makes sure we truncate the value before sending it to
the database.

In order to truncate intelligently, we first strip all tags from the
value. If the value is still too, big, then we truncate it.
Truncating html is prone to many errors and should never be done
blindly.